servicelocal/servicelocal/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.17
djangorestframework==3.8.2
idna==2.7
Markdown==3.0.1
mysqlclient==1.3.12
pip==18.1
pytz==2018.7
requests==2.20.1
setuptools==39.0.1
urllib3==1.24.1
